Fixed authentication/authorisation mixup + a couple little things missed in last commit.
